Well that was a stupid bug.
For some reason, instead of using (half_x, half_y) to define the new
size of the image, I used the coordinates of the current quarter.

BUT... THIS WORKED.

...because I just happened to be drawing the image at the lower right
quarter, where the starting coordinates just happened to be
(half_x, half_y).

Anyway something happened overnight and it crashed, but it's fixed now.